Add Level 2 Gameplay
Level 2 gameplay is pretty much the same as level one except that as you
continue it increases the number of lights in the sequence.

Level 1 - Always only 1 light
Level 2 - Every 3 turns add one extra light to a maximum of 10.

I probably should have started with the logic for Level 2 since Level 1
is the same with just the max number in the sequence limited to 1.

So I removed all the level 1 code since it is not needed.